requirements.txt
numpy==2.3.5
pillow==12.0.0
PyQt6==6.10.1
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.3
